Tender and Weapon
Part of prior commits, this is targeting the Tender and Weapon objects.
Work done on attack animations/line tracing/dealing damage logic for Tender.
Updating Gameplay Tag idea with a Tagging System that can be added to any actor. Idea is the same, but less copy/pasting logic and reworking multiple places.
Adjusted sockets on shovel for line traces